More alternatives to update item meta
