+++ src/test/java/net/minestom/server/item/ItemMetaTest.java
@@ -1,11 +1,28 @@
 package net.minestom.server.item;
 
 import net.minestom.server.item.metadata.BundleMeta;
+import org.jglrxavpok.hephaistos.nbt.NBT;
 import org.junit.jupiter.api.Test;
 
+import java.util.Map;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class ItemMetaTest {
+    @Test
+    public void defaultMeta() {
+        var item = ItemStack.builder(Material.BUNDLE).build();
+        assertNotNull(item.meta());
+    }
+
+    @Test
+    public void fromNBT() {
+        var compound = NBT.Compound(Map.of("value", NBT.Int(5)));
+        var item = ItemStack.builder(Material.BUNDLE).meta(compound).build();
+        assertEquals(compound, item.meta().toNBT());
+    }
+
     @Test
     public void bundle() {
         var item = ItemStack.builder(Material.BUNDLE)
