+++ src/test/java/net/minestom/server/item/ItemTest.java
@@ -57,20 +57,6 @@ public class ItemTest {
     }
 
     @Test
-    public void testAir() {
-        var item = ItemStack.of(Material.DIAMOND_SWORD);
-        assertFalse(item.isAir());
-        assertTrue(ItemStack.AIR.isAir());
-        var emptyItem = item.withAmount(0);
-        assertTrue(emptyItem.isAir());
-        assertEquals(emptyItem, ItemStack.AIR, "AIR item can be compared to empty item");
-        assertSame(emptyItem, ItemStack.AIR, "AIR item identity can be compared to empty item");
-
-        assertSame(ItemStack.AIR, ItemStack.fromNBT(Material.DIAMOND, null, 0));
-        assertSame(ItemStack.AIR, ItemStack.builder(Material.DIAMOND).amount(0).build());
-    }
-
-    @Test
     public void testItemNbt() {
         var itemNbt = createItem().toItemNBT();
         assertEquals(itemNbt.getString("id"), createItem().material().name(), "id string should be the material name");
